+++ src/lib/events.js
@@ -9,7 +9,7 @@
 
 'use strict';
 
-/* global $:false */
+/* global jQuery:false */
 
 var EventEmitter = require('events').EventEmitter;
 
@@ -53,8 +53,8 @@ var Events = {
          * events.
          */
         plotObj.emit = function(event, data) {
-            if(typeof $ !== 'undefined') {
-                $(plotObj).trigger(event, data);
+            if(typeof jQuery !== 'undefined') {
+                jQuery(plotObj).trigger(event, data);
             }
 
             ev.emit(event, data);
@@ -76,8 +76,8 @@ var Events = {
          * If Jquery exists run all its handlers for this event and
          * collect the return value of the LAST handler function
          */
-        if(typeof $ !== 'undefined') {
-            jQueryHandlerValue = $(plotObj).triggerHandler(event, data);
+        if(typeof jQuery !== 'undefined') {
+            jQueryHandlerValue = jQuery(plotObj).triggerHandler(event, data);
         }
 
         /*
@@ -114,7 +114,19 @@ var Events = {
          */
         return jQueryHandlerValue !== undefined ? jQueryHandlerValue :
             nodeEventHandlerValue;
+    },
+
+    purge: function(plotObj) {
+        delete plotObj._ev;
+        delete plotObj.on;
+        delete plotObj.once;
+        delete plotObj.removeListener;
+        delete plotObj.removeAllListeners;
+        delete plotObj.emit;
+
+        return plotObj;
     }
+
 };
 
 module.exports = Events;